% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdplearning-package.r
\docType{data}
\name{bluefin_tuna}
\alias{bluefin_tuna}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
Carl Boettiger
}
\keyword{data}
